+++ nudebomb.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 # Run nudebomb in development
 set -euo pipefail
-uv run ./nudebomb.py "$@"
+uv run nudebomb "$@"